feat: refactor generate text / object actions
